[Cherry-pick]Move etcd watch related code into eventmanager
PR: #21792
