+++ internal/datanode/data_node_test.go
@@ -18,7 +18,6 @@ package datanode
 
 import (
 	"context"
-	"fmt"
 	"math/rand"
 	"os"
 	"strconv"
@@ -26,13 +25,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/golang/protobuf/proto"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
 
 	"github.com/milvus-io/milvus-proto/go-api/v2/commonpb"
 	"github.com/milvus-io/milvus-proto/go-api/v2/schemapb"
-	etcdkv "github.com/milvus-io/milvus/internal/kv/etcd"
 	"github.com/milvus-io/milvus/internal/proto/datapb"
 	"github.com/milvus-io/milvus/internal/storage"
 	"github.com/milvus-io/milvus/internal/types"
@@ -189,7 +186,7 @@ func TestDataNode(t *testing.T) {
 	t.Run("Test BackGroundGC", func(t *testing.T) {
 		vchanNameCh := make(chan string)
 		node.clearSignal = vchanNameCh
-		node.wg.Add(1)
+		node.stopWaiter.Add(1)
 		go node.BackGroundGC(vchanNameCh)
 
 		testDataSyncs := []struct {
@@ -213,263 +210,5 @@ func TestDataNode(t *testing.T) {
 			}
 			return true
 		}, 2*time.Second, 10*time.Millisecond)
-	})
-
-}
-
-func TestWatchChannel(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	node := newIDLEDataNodeMock(ctx, schemapb.DataType_Int64)
-	etcdCli, err := etcd.GetEtcdClient(
-		Params.EtcdCfg.UseEmbedEtcd.GetAsBool(),
-		Params.EtcdCfg.EtcdUseSSL.GetAsBool(),
-		Params.EtcdCfg.Endpoints.GetAsStrings(),
-		Params.EtcdCfg.EtcdTLSCert.GetValue(),
-		Params.EtcdCfg.EtcdTLSKey.GetValue(),
-		Params.EtcdCfg.EtcdTLSCACert.GetValue(),
-		Params.EtcdCfg.EtcdTLSMinVersion.GetValue())
-	assert.NoError(t, err)
-	defer etcdCli.Close()
-	node.SetEtcdClient(etcdCli)
-	err = node.Init()
-	assert.NoError(t, err)
-	err = node.Start()
-	assert.NoError(t, err)
-	defer node.Stop()
-	err = node.Register()
-	assert.NoError(t, err)
-
-	defer cancel()
-
-	t.Run("test watch channel", func(t *testing.T) {
-		kv := etcdkv.NewEtcdKV(etcdCli, Params.EtcdCfg.MetaRootPath.GetValue())
-		oldInvalidCh := "datanode-etcd-test-by-dev-rootcoord-dml-channel-invalid"
-		path := fmt.Sprintf("%s/%d/%s", Params.CommonCfg.DataCoordWatchSubPath.GetValue(), paramtable.GetNodeID(), oldInvalidCh)
-		err = kv.Save(path, string([]byte{23}))
-		assert.NoError(t, err)
-
-		ch := fmt.Sprintf("datanode-etcd-test-by-dev-rootcoord-dml-channel_%d", rand.Int31())
-		path = fmt.Sprintf("%s/%d/%s", Params.CommonCfg.DataCoordWatchSubPath.GetValue(), paramtable.GetNodeID(), ch)
-
-		vchan := &datapb.VchannelInfo{
-			CollectionID:        1,
-			ChannelName:         ch,
-			UnflushedSegmentIds: []int64{},
-		}
-		info := &datapb.ChannelWatchInfo{
-			State: datapb.ChannelWatchState_ToWatch,
-			Vchan: vchan,
-		}
-		val, err := proto.Marshal(info)
-		assert.NoError(t, err)
-		err = kv.Save(path, string(val))
-		assert.NoError(t, err)
-
-		assert.Eventually(t, func() bool {
-			exist := node.flowgraphManager.exist(ch)
-			if !exist {
-				return false
-			}
-			bs, err := kv.LoadBytes(fmt.Sprintf("%s/%d/%s", Params.CommonCfg.DataCoordWatchSubPath.GetValue(), paramtable.GetNodeID(), ch))
-			if err != nil {
-				return false
-			}
-			watchInfo := &datapb.ChannelWatchInfo{}
-			err = proto.Unmarshal(bs, watchInfo)
-			if err != nil {
-				return false
-			}
-			return watchInfo.GetState() == datapb.ChannelWatchState_WatchSuccess
-		}, 3*time.Second, 100*time.Millisecond)
-
-		err = kv.RemoveWithPrefix(fmt.Sprintf("%s/%d", Params.CommonCfg.DataCoordWatchSubPath.GetValue(), paramtable.GetNodeID()))
-		assert.NoError(t, err)
-
-		assert.Eventually(t, func() bool {
-			exist := node.flowgraphManager.exist(ch)
-			return !exist
-		}, 3*time.Second, 100*time.Millisecond)
-	})
-
-	t.Run("Test release channel", func(t *testing.T) {
-		kv := etcdkv.NewEtcdKV(etcdCli, Params.EtcdCfg.MetaRootPath.GetValue())
-		oldInvalidCh := "datanode-etcd-test-by-dev-rootcoord-dml-channel-invalid"
-		path := fmt.Sprintf("%s/%d/%s", Params.CommonCfg.DataCoordWatchSubPath.GetValue(), paramtable.GetNodeID(), oldInvalidCh)
-		err = kv.Save(path, string([]byte{23}))
-		assert.NoError(t, err)
-
-		ch := fmt.Sprintf("datanode-etcd-test-by-dev-rootcoord-dml-channel_%d", rand.Int31())
-		path = fmt.Sprintf("%s/%d/%s", Params.CommonCfg.DataCoordWatchSubPath.GetValue(), paramtable.GetNodeID(), ch)
-		c := make(chan struct{})
-		go func() {
-			ec := kv.WatchWithPrefix(fmt.Sprintf("%s/%d", Params.CommonCfg.DataCoordWatchSubPath.GetValue(), paramtable.GetNodeID()))
-			c <- struct{}{}
-			cnt := 0
-			for {
-				evt := <-ec
-				for _, event := range evt.Events {
-					if strings.Contains(string(event.Kv.Key), ch) {
-						cnt++
-					}
-				}
-				if cnt >= 2 {
-					break
-				}
-			}
-			c <- struct{}{}
-		}()
-		// wait for check goroutine start Watch
-		<-c
-
-		vchan := &datapb.VchannelInfo{
-			CollectionID:        1,
-			ChannelName:         ch,
-			UnflushedSegmentIds: []int64{},
-		}
-		info := &datapb.ChannelWatchInfo{
-			State: datapb.ChannelWatchState_ToRelease,
-			Vchan: vchan,
-		}
-		val, err := proto.Marshal(info)
-		assert.NoError(t, err)
-		err = kv.Save(path, string(val))
-		assert.NoError(t, err)
-
-		// wait for check goroutine received 2 events
-		<-c
-		exist := node.flowgraphManager.exist(ch)
-		assert.False(t, exist)
-
-		err = kv.RemoveWithPrefix(fmt.Sprintf("%s/%d", Params.CommonCfg.DataCoordWatchSubPath.GetValue(), paramtable.GetNodeID()))
-		assert.NoError(t, err)
-		//TODO there is not way to sync Release done, use sleep for now
-		time.Sleep(100 * time.Millisecond)
-
-		exist = node.flowgraphManager.exist(ch)
-		assert.False(t, exist)
-
-	})
-
-	t.Run("handle watch info failed", func(t *testing.T) {
-		e := &event{
-			eventType: putEventType,
-		}
-
-		node.handleWatchInfo(e, "test1", []byte{23})
-
-		exist := node.flowgraphManager.exist("test1")
-		assert.False(t, exist)
-
-		info := datapb.ChannelWatchInfo{
-			Vchan: nil,
-			State: datapb.ChannelWatchState_Uncomplete,
-		}
-		bs, err := proto.Marshal(&info)
-		assert.NoError(t, err)
-		node.handleWatchInfo(e, "test2", bs)
-
-		exist = node.flowgraphManager.exist("test2")
-		assert.False(t, exist)
-
-		chPut := make(chan struct{}, 1)
-		chDel := make(chan struct{}, 1)
-
-		ch := fmt.Sprintf("datanode-etcd-test-by-dev-rootcoord-dml-channel_%d", rand.Int31())
-		m := newChannelEventManager(
-			func(info *datapb.ChannelWatchInfo, version int64) error {
-				r := node.handlePutEvent(info, version)
-				chPut <- struct{}{}
-				return r
-			},
-			func(vChan string) {
-				node.handleDeleteEvent(vChan)
-				chDel <- struct{}{}
-			}, time.Millisecond*100,
-		)
-		node.eventManagerMap.Insert(ch, m)
-		m.Run()
-		defer m.Close()
-
-		info = datapb.ChannelWatchInfo{
-			Vchan: &datapb.VchannelInfo{ChannelName: ch},
-			State: datapb.ChannelWatchState_Uncomplete,
-		}
-		bs, err = proto.Marshal(&info)
-		assert.NoError(t, err)
-
-		msFactory := node.factory
-		defer func() { node.factory = msFactory }()
-
-		// todo review the UT logic
-		// As we remove timetick channel logic, flow_graph_insert_buffer_node no longer depend on MessageStreamFactory
-		// so data_sync_service can be created. this assert becomes true
-		node.factory = &FailMessageStreamFactory{}
-		node.handleWatchInfo(e, ch, bs)
-		<-chPut
-		exist = node.flowgraphManager.exist(ch)
-		assert.True(t, exist)
-	})
-
-	t.Run("handle watchinfo out of date", func(t *testing.T) {
-		chPut := make(chan struct{}, 1)
-		chDel := make(chan struct{}, 1)
-		// inject eventManager
-		ch := fmt.Sprintf("datanode-etcd-test-by-dev-rootcoord-dml-channel_%d", rand.Int31())
-		m := newChannelEventManager(
-			func(info *datapb.ChannelWatchInfo, version int64) error {
-				r := node.handlePutEvent(info, version)
-				chPut <- struct{}{}
-				return r
-			},
-			func(vChan string) {
-				node.handleDeleteEvent(vChan)
-				chDel <- struct{}{}
-			}, time.Millisecond*100,
-		)
-		node.eventManagerMap.Insert(ch, m)
-		m.Run()
-		defer m.Close()
-		e := &event{
-			eventType: putEventType,
-			version:   10000,
-		}
-
-		info := datapb.ChannelWatchInfo{
-			Vchan: &datapb.VchannelInfo{ChannelName: ch},
-			State: datapb.ChannelWatchState_Uncomplete,
-		}
-		bs, err := proto.Marshal(&info)
-		assert.NoError(t, err)
-
-		node.handleWatchInfo(e, ch, bs)
-		<-chPut
-		exist := node.flowgraphManager.exist("test3")
-		assert.False(t, exist)
-	})
-
-	t.Run("handle watchinfo compatibility", func(t *testing.T) {
-		info := datapb.ChannelWatchInfo{
-			Vchan: &datapb.VchannelInfo{
-				CollectionID:        1,
-				ChannelName:         "delta-channel1",
-				UnflushedSegments:   []*datapb.SegmentInfo{{ID: 1}},
-				FlushedSegments:     []*datapb.SegmentInfo{{ID: 2}},
-				DroppedSegments:     []*datapb.SegmentInfo{{ID: 3}},
-				UnflushedSegmentIds: []int64{1},
-			},
-			State: datapb.ChannelWatchState_Uncomplete,
-		}
-		bs, err := proto.Marshal(&info)
-		assert.NoError(t, err)
-
-		newWatchInfo, err := parsePutEventData(bs)
-		assert.NoError(t, err)
-
-		assert.Equal(t, []*datapb.SegmentInfo{}, newWatchInfo.GetVchan().GetUnflushedSegments())
-		assert.Equal(t, []*datapb.SegmentInfo{}, newWatchInfo.GetVchan().GetFlushedSegments())
-		assert.Equal(t, []*datapb.SegmentInfo{}, newWatchInfo.GetVchan().GetDroppedSegments())
-		assert.NotEmpty(t, newWatchInfo.GetVchan().GetUnflushedSegmentIds())
-		assert.NotEmpty(t, newWatchInfo.GetVchan().GetFlushedSegmentIds())
-		assert.NotEmpty(t, newWatchInfo.GetVchan().GetDroppedSegmentIds())
 	})
 }
